src/consumer/requirements.txt
kafka-python==2.0.2
boto3==1.34.0
pandas==2.1.4
python-dotenv==1.0.0
pyarrow==14.0.2 